+++ lib/class/Components/ScrubbingForward/index.d.ts
@@ -2,14 +2,20 @@ import { IConstructorBaseProps } from '../../../type';
 import BaseComponent from '../../BaseComponent';
 interface IConstructorProps extends IConstructorBaseProps {
 }
-declare class CurrentTime extends BaseComponent {
+declare class ScrubbingForward extends BaseComponent {
+    private counter;
+    private timerId;
+    private timerIdScrubbing;
+    private icon;
+    private text;
+    private ripple;
     constructor(props: IConstructorProps);
     render(): void;
     registerListener(): void;
     unregisterListener(): void;
-    update(value: number): void;
-    hide(): void;
+    handleContainerClick(event: MouseEvent | TouchEvent): void;
     show(): void;
+    hidden(): void;
 }
-export default CurrentTime;
+export default ScrubbingForward;
 //# sourceMappingURL=index.d.ts.map